Start controller submission work
This commit implements a new Submit() method to be used by the Connect API when
we're creating a new package. Most of the work is focused on the deposit of the
transfer into the shared directory which involves talking to the Archivematica
Storage Service using the ssclient package.

There's more work to do, I haven't tested the RPC yet. I've imported the latest
version of go.artefactual.dev/ssclient (v0.1.0).