Implement alternative worker engine
This commit introduces a new worker engine that runs alongside the existing
Gearman-based Python worker (MCPClient). The main goal is to provide more
flexibility and allow Go-based workers, including the option to embed them
directly inside the main CCP process. The main hypothesis is whether it's
possible to design a system that can be deployed as a single binary, without
the need for independently deployed MCPClient workers or a MySQL database,
and without sacrificing horizontal scalability - something that Archivematica
achives with difficulty and some limitations, and which a3m trades off
for simplicity. The fact that both types of workers can coexist means we can
refactor the code gradually, in small steps.

Can we build a compliant derivative of Archivematica that is as simple to deploy
in single-node environments - think Forgejo or Pydio — but is also truly
scalable at the same time?

In this new system, workers and the server communicate using a Connect/RPC API.
The service is called `WorkerService` and is defined with Protobuf at
`proto/archivematica/ccp/worker/v1beta1`. The server and all worker management
code are in `workhub` package. On the client side, the worker is implemented in
`worker` package. Additionally, `provisioner` optionally manages automatic
deployment of embedded workers by configuration. The workflow `controller` now
uses a `dispatcher` to send jobs, removing the dependency on MCPClient.
It performs as a simple router based on the client script name.

Currently, the new worker only supports the execution of `test_v0.0` jobs. When
CCP runs this script, it uses the function in `internal/worker/fn/test.go`. The
interface between these new functions and the worker still needs more work.
There are still some rough edges in different parts, and I will keep improving
them as I can. But overall, everything seems to work the same as before these
changes.